Fixes a bug where output event emitters were being shared.
Setting the output EventEmitter property on the prototype means that the last value set will be used by all instances of the Mock class. This is likely not the intended behavior and it means you cannot test output events for multiple component instances.

Moving this to inside the Mock constructor ensures that every mock instance gets its own EventEmitter.